requirements-dev.txt
asgiref==3.5.2
bleach==5.0.1
cachetools==5.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-allauth==0.51.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.1
django-heroku==0.3.1
django-markdownify==0.9.2
django-rest-auth==0.9.5
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.0
environ==1.0
flake8==3.9.0
fonttools==4.37.4
google-api-core==1.33.2
google-api-python-client==1.10.0
google-auth==2.13.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.1.0
httplib2==0.20.4
idna==2.10
importlib-metadata==3.7.3
kiwisolver==1.4.4
Markdown==3.2.2
matplotlib
mccabe==0.6.1
numpy
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==3.20.3
psycopg2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tinycss2==1.1.1
typing-extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.1
nested-multipart-parser==1.4.1
drf_yasg
pydantic==1.10.7